+++ packages/patrol/android/src/main/kotlin/pl/leancode/patrol/contracts/PatrolAppServiceClient.kt
@@ -19,6 +19,16 @@ class PatrolAppServiceClient(address: String, port: Int, private val timeout: Lo
         return json.fromJson(response, Contracts.ListDartTestsResponse::class.java)
     }
 
+    fun listDartLifecycleCallbacks(): Contracts.ListDartLifecycleCallbacksResponse {
+        val response = performRequest("listDartLifecycleCallbacks")
+        return json.fromJson(response, Contracts.ListDartLifecycleCallbacksResponse::class.java)
+    }
+
+    fun setLifecycleCallbacksState(request: Contracts.SetLifecycleCallbacksStateRequest): Contracts.Empty {
+        val response = performRequest("setLifecycleCallbacksState", json.toJson(request))
+        return json.fromJson(response, Contracts.Empty::class.java)
+    }
+
     fun runDartTest(request: Contracts.RunDartTestRequest): Contracts.RunDartTestResponse {
         val response = performRequest("runDartTest", json.toJson(request))
         return json.fromJson(response, Contracts.RunDartTestResponse::class.java)