+++ packages/patrol/darwin/Classes/AutomatorServer/NativeAutomatorServer.swift
@@ -41,6 +41,7 @@ protocol NativeAutomatorServer {
     func setLocationAccuracy(request: SetLocationAccuracyRequest) throws
     func debug() throws
     func markPatrolAppServiceReady() throws
+    func markLifecycleCallbackExecuted(request: MarkLifecycleCallbackExecutedRequest) throws
 }
 
 extension NativeAutomatorServer {
@@ -239,6 +240,12 @@ extension NativeAutomatorServer {
         try markPatrolAppServiceReady()
         return HTTPResponse(.ok)
     }
+
+    private func markLifecycleCallbackExecutedHandler(request: HTTPRequest) throws -> HTTPResponse {
+        let requestArg = try JSONDecoder().decode(MarkLifecycleCallbackExecutedRequest.self, from: request.body)
+        try markLifecycleCallbackExecuted(request: requestArg)
+        return HTTPResponse(.ok)
+    }
 }
 
 extension NativeAutomatorServer {
@@ -417,6 +424,11 @@ extension NativeAutomatorServer {
             request in handleRequest(
                 request: request,
                 handler: markPatrolAppServiceReadyHandler)
+        }
+        server.route(.POST, "markLifecycleCallbackExecuted") {
+            request in handleRequest(
+                request: request,
+                handler: markLifecycleCallbackExecutedHandler)
         }
     }
 }