+++ packages/patrol/lib/src/native/contracts/patrol_app_service_server.dart
@@ -18,6 +18,22 @@ abstract class PatrolAppServiceServer {
 
       final body = jsonEncode(result.toJson());
       return Response.ok(body);
+    } else if ('listDartLifecycleCallbacks' == request.url.path) {
+      final result = await listDartLifecycleCallbacks();
+
+      final body = jsonEncode(result.toJson());
+      return Response.ok(body);
+    } else if ('setLifecycleCallbacksState' == request.url.path) {
+      final stringContent = await request.readAsString(utf8);
+      final json = jsonDecode(stringContent);
+      final requestObj = SetLifecycleCallbacksStateRequest.fromJson(
+        json as Map<String, dynamic>,
+      );
+
+      final result = await setLifecycleCallbacksState(requestObj);
+
+      final body = jsonEncode(result.toJson());
+      return Response.ok(body);
     } else if ('runDartTest' == request.url.path) {
       final stringContent = await request.readAsString(utf8);
       final json = jsonDecode(stringContent);
@@ -34,5 +50,9 @@ abstract class PatrolAppServiceServer {
   }
 
   Future<ListDartTestsResponse> listDartTests();
+  Future<ListDartLifecycleCallbacksResponse> listDartLifecycleCallbacks();
+  Future<Empty> setLifecycleCallbacksState(
+    SetLifecycleCallbacksStateRequest request,
+  );
   Future<RunDartTestResponse> runDartTest(RunDartTestRequest request);
 }