+++ packages/patrol/CHANGELOG.md
@@ -4,6 +4,13 @@
 	vanilla `setUp()` and `tearDown()` but work with Patrol's native automation
 	and test bundling feature (#1967) (#1721)
 
+## 3.1.0-dev.1
+
+- Add `patrolSetUp()` and `patrolTearDown()` (#1721)
+- Add `patrolSetUpAll()` (#1751)
+
+This version requires version 2.4.0-dev.1 of `patrol_cli` package.
+
 ## 3.0.3
 
 - Implement `$.native.swipe()` on iOS (#1927)
@@ -36,6 +43,7 @@ Other changes:
 
 - Bump minimum supported Flutter version to 3.16
 - **BREAKING:**
+
   - Remove `bindingType` parameter from `patrolTest()` function. Now only
     `PatrolBinding` is used and it's automatically initialized (#1882)
   - Remove `nativeAutomation` parameter from `patrolTest()` function. Now it's